refactor stubs into separate file
